+++ owncloudApp/src/main/java/com/owncloud/android/ui/activity/ReceiveExternalFilesActivity.java
@@ -141,7 +141,6 @@ public class ReceiveExternalFilesActivity extends FileActivity
     private String mUploadPath;
     private OCFile mFile;
     private SortOptionsView mSortOptionsView;
-    private SearchView mSearchView;
 
     private View mEmptyListView;
     private ImageView mEmptyListImage;
@@ -178,6 +177,8 @@ public class ReceiveExternalFilesActivity extends FileActivity
     private boolean showHiddenFiles;
     private OCSharedPreferencesProvider sharedPreferencesProvider;
     private OCSpace personalSpace;
+
+    private boolean isInSpacesList;
 
     Pattern pattern = Pattern.compile("[/\\\\]");
 
@@ -236,7 +237,6 @@ public class ReceiveExternalFilesActivity extends FileActivity
                     updateToolbar(getString(R.string.uploader_top_message));
                 }
             } else { // OCIS Server
-
                 if (haveMultiAccount) { // Multi account
                     mListView = findViewById(android.R.id.list);
                     fragmentContainer = findViewById(R.id.fragment_container);
@@ -258,6 +258,7 @@ public class ReceiveExternalFilesActivity extends FileActivity
     }
 
     private void showListOfFiles() {
+        isInSpacesList = false;
         fragmentContainer = findViewById(R.id.fragment_container);
         mListView = findViewById(android.R.id.list);
         fragmentContainer.setVisibility(View.GONE);
@@ -292,6 +293,7 @@ public class ReceiveExternalFilesActivity extends FileActivity
         uploaderButton = findViewById(R.id.uploader_choose_folder);
         uploaderButton.setVisibility(View.GONE);
         updateToolbar(getString(R.string.choose_upload_space));
+        isInSpacesList = true;
     }
 
     private void initPickerListener() {
@@ -753,12 +755,14 @@ public class ReceiveExternalFilesActivity extends FileActivity
     public boolean onCreateOptionsMenu(Menu menu) {
         MenuInflater inflater = getMenuInflater();
         inflater.inflate(R.menu.main_menu, menu);
-
-        mSearchView = (SearchView) menu.findItem(R.id.action_search).getActionView();
+        MenuItem menuItem = menu.findItem(R.id.action_search);
+        SearchView mSearchView = (SearchView) menuItem.getActionView();
         mSearchView.setMaxWidth(Integer.MAX_VALUE);
         mSearchView.setQueryHint(getResources().getString(R.string.actionbar_search));
         mSearchView.setOnQueryTextListener(this);
-
+        if (isInSpacesList) {
+            menuItem.setVisible(false);
+        }
         menu.removeItem(menu.findItem(R.id.action_share_current_folder).getItemId());
 
         return true;